protection on server actions

+++ src/server/db/actions/PostActions.ts
@@ -182,8 +182,12 @@ function postPopulationPipeline({
                   $match: {
                     $or: [
                       // No childBirthdates - include these posts
-                      ...(age.maxAge === 100 ? [{ "ageData.childBirthdates": { $exists: false } },
-                        { "ageData.childBirthdates": { $size: 0 } }] : []),
+                      ...(age.maxAge === 100
+                        ? [
+                            { "ageData.childBirthdates": { $exists: false } },
+                            { "ageData.childBirthdates": { $size: 0 } },
+                          ]
+                        : []),
 
                       // Has at least one child in the age range
                       {
@@ -342,12 +346,19 @@ function postPopulationPipeline({
 export async function createPost(post: PostInput): Promise<Post> {
   await dbConnect();
 
+  const currentUser = await getAuthenticatedUser();
+  if (!currentUser || currentUser.isBanned) {
+    throw new Error("User does not have access");
+  }
+
   const profanities = await getAllProfanities();
   const profanityWords = profanities.map((profanity) => profanity.name);
 
   const contentProfanities = containsProfanity(post.content, profanityWords);
   const titleProfanities = containsProfanity(post.title, profanityWords);
-  const uniqueProfanities = Array.from(new Set([...contentProfanities, ...titleProfanities]));
+  const uniqueProfanities = Array.from(
+    new Set([...contentProfanities, ...titleProfanities]),
+  );
 
   const isFlagged = uniqueProfanities.length > 0;
 
@@ -372,7 +383,6 @@ export async function createPost(post: PostInput): Promise<Post> {
   return createdPost.toObject();
 }
 
-
 /**
  * Validates a post to determine if it will be flagged based on its content and title.
  * @param post - The post input data.
@@ -386,11 +396,12 @@ export async function validatePost(post: PostInput): Promise<string[]> {
 
   const contentProfanities = containsProfanity(post.content, profanityWords);
   const titleProfanities = containsProfanity(post.title, profanityWords);
-  const uniqueProfanities = Array.from(new Set([...contentProfanities, ...titleProfanities]));
+  const uniqueProfanities = Array.from(
+    new Set([...contentProfanities, ...titleProfanities]),
+  );
 
   return uniqueProfanities;
 }
-
 
 /**
  * Pins a post in the database.
@@ -455,7 +466,7 @@ export async function unpinPost(
   await dbConnect();
   const currentUser = await getAuthenticatedUser();
   if (!currentUser?.isAdmin) {
-    throw new Error("This edit can only be made by an admin");
+    throw new Error("Only admins can unpin posts");
   }
 
   if (!mongoose.Types.ObjectId.isValid(postId)) {
@@ -490,6 +501,11 @@ export async function getPopulatedPinnedPosts(
   authUserId: string,
 ): Promise<PostAggregationResult> {
   await dbConnect();
+
+  const currentUser = await getAuthenticatedUser();
+  if (!currentUser) {
+    throw new Error("User does not have access");
+  }
 
   const pipeline = [
     { $match: { isPinned: true, isPrivate: false } },
@@ -535,6 +551,11 @@ export async function getPopulatedPosts(
 ): Promise<PostAggregationResult> {
   await dbConnect();
 
+  const currentUser = await getAuthenticatedUser();
+  if (!currentUser) {
+    throw new Error("User does not have access");
+  }
+
   const postsInfo = await PostModel.aggregate(
     postPopulationPipeline({
       authUserId,
@@ -567,6 +588,11 @@ export async function getPopulatedUserPosts(
   isAdmin?: boolean,
 ): Promise<PopulatedPost[]> {
   await dbConnect();
+
+  const currentUser = await getAuthenticatedUser();
+  if (!currentUser) {
+    throw new Error("User does not have access");
+  }
 
   const postsShown =
     currUserId === userId || isAdmin
@@ -603,6 +629,11 @@ export async function getPopulatedPost(
     throw new Error("Invalid post ID");
   }
 
+  const currentUser = await getAuthenticatedUser();
+  if (!currentUser) {
+    throw new Error("User does not have access");
+  }
+
   const aggregationResult = await PostModel.aggregate(
     postPopulationPipeline({
       authUserId,
@@ -613,6 +644,7 @@ export async function getPopulatedPost(
   );
   const post = aggregationResult[0].posts[0];
   if (!post) {
+    console.log("here: " + id);
     throw new Error("Post not found");
   }
   return post;
@@ -635,6 +667,16 @@ export async function editPost(
     throw new Error("Invalid post ID");
   }
 
+  const currentUser = await getAuthenticatedUser();
+  const postInfo = await PostModel.findById(id);
+  if (
+    !currentUser ||
+    (!currentUser.isAdmin && currentUser._id !== postInfo.author?.toString()) ||
+    currentUser.isBanned
+  ) {
+    throw new Error("User does not have access");
+  }
+
   const validatedPost = editPostSchema.parse(post);
   if (post.editedByAdmin !== undefined) {
     validatedPost.editedByAdmin = post.editedByAdmin;
@@ -642,7 +684,6 @@ export async function editPost(
   const updatedPost = await PostModel.findByIdAndUpdate(id, validatedPost, {
     new: true,
   });
-  console.log(updatedPost);
   if (!updatedPost) {
     throw new Error("Post not found");
   }
@@ -666,6 +707,14 @@ export async function deletePost(
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
     throw new Error("Invalid post ID");
+  }
+
+  const currentUser = await getAuthenticatedUser();
+  if (
+    !currentUser ||
+    (!currentUser.isAdmin && currentUser._id !== authUserId)
+  ) {
+    throw new Error("User does not have access");
   }
 
   const updatedPost = await PostModel.findByIdAndUpdate(id, {
@@ -715,6 +764,11 @@ export async function createPostSave(
     throw new Error("Invalid user ID or post ID");
   }
 
+  const currentUser = await getAuthenticatedUser();
+  if (!currentUser || (!currentUser.isAdmin && currentUser._id !== userId)) {
+    throw new Error("User does not have access");
+  }
+
   const existingSave = await PostSaveModel.findOne({
     user: userId,
     post: postId,
@@ -744,6 +798,11 @@ export async function getSavedPosts(userId: string): Promise<Post[]> {
     throw new Error("Invalid user ID");
   }
 
+  const currentUser = await getAuthenticatedUser();
+  if (!currentUser) {
+    throw new Error("User does not have access");
+  }
+
   const savedPosts = await PostSaveModel.find({ user: userId })
     .sort({ date: -1 })
     .populate("post")
@@ -765,6 +824,11 @@ export async function getPopulatedSavedPosts(
 
   if (!mongoose.Types.ObjectId.isValid(userId)) {
     throw new Error("Invalid user ID");
+  }
+
+  const currentUser = await getAuthenticatedUser();
+  if (!currentUser) {
+    throw new Error("User does not have access");
   }
 
   const pipeline: mongoose.PipelineStage[] = [
@@ -812,6 +876,11 @@ export async function deletePostSave(
     throw new Error("Invalid user ID or post ID");
   }
 
+  const currentUser = await getAuthenticatedUser();
+  if (!currentUser || (!currentUser.isAdmin && currentUser._id !== userId)) {
+    throw new Error("User does not have access");
+  }
+
   const deletedSave = await PostSaveModel.findOneAndDelete({
     user: userId,
     post: postId,
@@ -840,6 +909,11 @@ export async function createPostLike(
     !mongoose.Types.ObjectId.isValid(postId)
   ) {
     throw new Error("Invalid user ID or post ID");
+  }
+
+  const currentUser = await getAuthenticatedUser();
+  if (!currentUser || (!currentUser.isAdmin && currentUser._id !== userId)) {
+    throw new Error("User does not have access");
   }
 
   const session = await PostModel.startSession();
@@ -900,6 +974,11 @@ export async function deletePostLike(
     throw new Error("Invalid user ID or post ID");
   }
 
+  const currentUser = await getAuthenticatedUser();
+  if (!currentUser || (!currentUser.isAdmin && currentUser._id !== userId)) {
+    throw new Error("User does not have access");
+  }
+
   const session = await PostModel.startSession();
   session.startTransaction();
 
@@ -939,6 +1018,10 @@ export async function deletePostLike(
  */
 export async function hasFlaggedPosts(): Promise<boolean> {
   await dbConnect();
+  const currentUser = await getAuthenticatedUser();
+  if (!currentUser || !currentUser.isAdmin) {
+    throw new Error("User does not access");
+  }
   const flaggedPostsCount = await PostModel.countDocuments({
     isFlagged: true,
     isDeleted: false,
